Fix pg12 compatibility wrt. 7d872c91a3f
This commit introduced a new root->append_rel_array that's setup just
before we get a change to fill in the necessary data in root->append_rel_list.

For now, we call setup_append_rel_array() (and free previously palloc'd
memory if needed) for each hypo_injectHypotheticalPartitioning() call,
but this is a waste of cycles, we'll need to call it ideally only once
per query execution.