Reintroduce childrel expanding order for pg10
The partitions has to be sorted by bound order, otherwise many things
can break.

This commit also introduce specific regression tests for pg10, for which
the duplucate sanity check would not exhibit the expected error without
the expanded order.

Some of these new regression tests still fails, due to minor change in
the join or partition order, we'll see if we can fix the behavior or change
the expected results.